asset reconciliation daemon
