+++ python_modules/dagster/dagster/_core/definitions/asset_layer.py
@@ -35,9 +35,9 @@ from .node_definition import NodeDefinition
 from .resource_definition import ResourceDefinition
 
 if TYPE_CHECKING:
-    from dagster._core.definitions.partition_mapping import PartitionMapping
     from dagster._core.definitions.assets import AssetsDefinition, SourceAsset
     from dagster._core.definitions.job_definition import JobDefinition
+    from dagster._core.definitions.partition_mapping import PartitionMapping
     from dagster._core.definitions.resolved_asset_defs import ResolvedAssetDependencies
     from dagster._core.execution.context.output import OutputContext
